pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "localstack-sample-tester"
version = "0.1.0"
description = "Automated compatibility scanner for AWS/Azure sample apps against LocalStack"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "PyGithub>=2.1",
    "docker>=7.0",
    "jinja2>=3.1",
    "ruamel.yaml>=0.18",
    "requests>=2.31",
]

[project.scripts]
scanner = "scanner.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/scanner"]

[tool.hatch.build.targets.editable]
path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "integration: marks tests requiring real LocalStack instance",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.15.6",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = [
    "E501",   # Long lines — Click help strings and mock chains are structurally long
    "UP042",  # StrEnum migration — out of scope for initial setup
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B"]
